requirements.txt
streamlit==1.33.0
langchain==0.1.20
langchain-community==0.0.38
chromadb==0.4.24
sentence-transformers==2.7.0
python-dotenv==1.0.1
langchain_groq==0.1.3
langsmith==0.1.63
